Fixed Remaining Issues
1. Added sorting after swapping.
2. That fixed all the "Straights". Now you can swap until you build a
"Straight" and it'll work. Drawn "Straights" still worked if I'm not
mistaken.
3. Added "kept" when choosing to keep all the cards in hand. (Or
choosing to discard none of the cards :P )
4. I'm still weirded out by the memory leak because I just can't find
it. I added a file.close() for both of the filestreams I'm using to the
exit function.
5. Scratch that, the problem was that when I switched the deck with the
discard pile, the original deck pointer never got deleted. So now I just
delete the deck and pointed to NULL before switching it to the dump.